% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_heatmap}
\alias{plot_heatmap}
\alias{plot_heatmap.matrix}
\alias{plot_heatmap.data.frame}
\title{Plot a heatmap of a trn.}
\usage{
plot_heatmap(x, guide = TRUE, limits = NULL)

\method{plot_heatmap}{matrix}(x, guide = TRUE, limits = NULL)

\method{plot_heatmap}{data.frame}(x, guide = TRUE, limits = NULL)
}
\arguments{
\item{x}{a TRN object.}

\item{guide}{logical; whether to plot the guide.}

\item{limits}{set limits to scale_fill_gradient2().}
}
\description{
Plot a heatmap of a trn.
}
