% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_trn}
\alias{plot_trn}
\title{Plot transcriptional regulatory networks.}
\usage{
plot_trn(x, ...)
}
\arguments{
\item{x}{a network object with attributes.}
}
\description{
Plot transcriptional regulatory networks.
}
