% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_trn_graph}
\alias{plot_trn_graph}
\title{plot_trn_graph}
\usage{
plot_trn_graph(
  x,
  cutoff = NULL,
  layout = "nicely",
  color.by = NULL,
  seed = 42,
  ...
)
}
\arguments{
\item{x}{}
}
\description{
plot_trn_graph
}
