% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.R
\name{trm_lm}
\alias{trm_lm}
\title{Fits a linear model to all genes using a matrix of regulators' activities as predictors.}
\usage{
trm_lm(x, y)
}
\arguments{
\item{x}{matrix of regulators' activities.}

\item{y}{matrix of genes' expression levels.}
}
\description{
Fits a linear model to all genes using a matrix of regulators' activities as predictors.
}
