% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.R
\name{trm_lm_by_group}
\alias{trm_lm_by_group}
\title{Fits a linear model to all genes using a matrix of regulators' activities as predictors
and a grouping variable.}
\usage{
trm_lm_by_group(x, y, group)
}
\arguments{
\item{x}{matrix of regulators' activities.}

\item{y}{matrix of genes' expression levels.}

\item{group}{grouping variable.}
}
\description{
Fits a linear model to all genes using a matrix of regulators' activities as predictors
and a grouping variable.
}
