% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{trn_compare}
\alias{trn_compare}
\title{Computes percentage of coherent directions between two TRN models.}
\usage{
trn_compare(x, y, direction = "all")
}
\arguments{
\item{x}{a TRN object (truth).}

\item{y}{a TRN object (predicted).}

\item{direction}{whether to use all edges, empty edges, positive edges or negative edges}
}
\description{
Computes percentage of coherent directions between two TRN models.
}
