% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.R
\name{trn_filter_glmnet}
\alias{trn_filter_glmnet}
\title{Filter a TRN at a given lambda for networks fitted using penalized regression.}
\usage{
trn_filter_glmnet(x, s = 0.01)
}
\arguments{
\item{x}{a TRN model.}

\item{s}{value of penalty parameter.}
}
\description{
Filter a TRN at a given lambda for networks fitted using penalized regression.
}
