% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{trn_fit}
\alias{trn_fit}
\alias{trn_fit.Seurat}
\title{trn_fit}
\usage{
trn_fit(x, db, ...)

\method{trn_fit}{Seurat}(x, db, assay = NULL, slot = "data", scale = TRUE, filter = TRUE, ...)
}
\arguments{
\item{x}{dataset with gene expression levels.}

\item{db}{list of transcription factors.}

\item{...}{arguments passed down to methods.}

\item{assay}{assay for Seurat objects.}

\item{slot}{slot for Seurat objects.}

\item{scale}{whether to scale the data (only when slot is not scale.data).}

\item{filter}{whether to filter out genes with zero standard deviation.}
}
\description{
Fit a TRN model.
}
