% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.R
\name{trn_glmnet}
\alias{trn_glmnet}
\title{Fits a penalized linear model to all genes using a matrix of regulators' activities as predictors.}
\usage{
trn_glmnet(x, y)
}
\arguments{
\item{x}{matrix of regulators' activities.}

\item{y}{matrix of genes' expression levels.}
}
\description{
Fits a penalized linear model to all genes using a matrix of regulators' activities as predictors.
}
