% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.R
\name{trn_glmnet_by_group}
\alias{trn_glmnet_by_group}
\title{trn_glmnet_by_group}
\usage{
trn_glmnet_by_group(
  x,
  y,
  group,
  alpha = 0.5,
  s = 0.01,
  low.memory = FALSE,
  mc.cores = 1
)
}
\arguments{
\item{x}{matrix of regulators' activities.}

\item{y}{matrix of genes' expression levels.}

\item{group}{grouping variable.}

\item{alpha}{control balance between lasso (alpha = 1) and ridge (alpha = 0) penalties.}

\item{s}{value of penalty parameter.}

\item{low.memory}{whether to save memory by returning coefficients only.}

\item{mc.cores}{cores to use for parallel processing.}
}
\description{
trn_glmnet_by_group
}
