% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{trn_rmsd}
\alias{trn_rmsd}
\title{Computes RMSD between two TRN models.}
\usage{
trn_rmsd(x, y)
}
\arguments{
\item{x}{a TRN object.}

\item{y}{a TRN object.}
}
\description{
Takes a real and predicted model and computes the root mean squared distance.
}
