% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{trntools}
\alias{trntools}
\alias{trntools-package}
\title{Tools for the analysis of transcriptional regulatory networks}
\description{
Tools for the analysis of transcriptional regulatory networks
}
