% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_heatmap}
\alias{plot_heatmap}
\alias{plot_heatmap.matrix}
\title{Plot a heatmap of a trn.}
\usage{
plot_heatmap(x)

\method{plot_heatmap}{matrix}(x)
}
\arguments{
\item{x}{a TRN object.}
}
\description{
Plot a heatmap of a trn.
}
