% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_trm}
\alias{plot_trm}
\alias{plot_trm.data.frame}
\alias{plot_trm.tbl_graph}
\title{Plot TRMs.}
\usage{
plot_trm(x, ...)

\method{plot_trm}{data.frame}(x, node.size = 10, label.size = 3,
  regulator.color = "palevioletred1", target.color = "turquoise3")

\method{plot_trm}{tbl_graph}(x, layout = "nicely", ...)
}
\arguments{
\item{x}{a TRM object.}
}
\description{
Plot TRMs.
}
