% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{trn_filter}
\alias{trn_filter}
\title{Filter a TRN at a given FDR.}
\usage{
trn_filter(x, fdr = 0.01)
}
\arguments{
\item{x}{a TRN model.}

\item{fdr}{FDR level.}
}
\description{
Filter a TRN at a given FDR.
}
